Provide documentation for Monitoring/Metrics
We wrote a LibreChat metrics exporter for Prometheus a while ago and
have been running it on our production LibreChat for a while to monitor
how the tool is being used and to alert us if there are any problems.

This patch adds a simple documentation (because deploying is actually
pretty simple) of how to set-up the metrics exporter for your LibreChat.

We are hoping this will help others as well.